refactoring for api blueprint continues
