+++ sspi_flask_app/api/core/collect.py
@@ -1,8 +1,8 @@
 from flask import Blueprint, Response
 from flask_login import login_required, current_user
-from ..source_utilities.sdg import collectSDGIndicatorData
-from ..source_utilities.iea import collectIEAData
-from ..dashboard import parse_json
+from ..datasource.sdg import collectSDGIndicatorData
+from ..datasource.iea import collectIEAData
+from .dashboard import parse_json
 from flask import redirect, url_for
 import datetime
 import requests
